Allow using sqlite for testing.  Set conftest.{BIND,DBNAME}.
Several changes: some postgres-specific DB ops are conditioned on
using postgres or provided with sqlite options, and, sadly, a set of
tests must be skipped because sqlalchemy's sqlite driver must have an
unpatched datetime.datetime for an isinstance() internally and I
can't figure out any other way to patch for the tests in
async/test_recurring.py

+++ chacra/tests/conftest.py
@@ -98,8 +98,9 @@ def connection(app, request):
     print("=" * 80)
     print("CREATING TEMPORARY DATABASE FOR TESTS")
     print("=" * 80)
-    subprocess.call(['dropdb', DBNAME])
-    subprocess.call(['createdb', DBNAME])
+    if BIND.startswith('postgresql'):
+        subprocess.call(['dropdb', DBNAME])
+        subprocess.call(['createdb', DBNAME])
 
     # Bind and create the database tables
     _db.clear()
@@ -154,14 +155,17 @@ def session(connection, request):
         conn = engine.connect()
         trans = conn.begin()
 
-        inspector = reflection.Inspector.from_engine(engine)
 
         # gather all data first before dropping anything.
         # some DBs lock after things have been dropped in
         # a transaction.
-        conn.execute("TRUNCATE TABLE %s RESTART IDENTITY CASCADE" % (
-            ', '.join(inspector.get_table_names())
-        ))
+        if BIND.startswith('postgresql'):
+            conn.execute("TRUNCATE TABLE %s RESTART IDENTITY CASCADE" % (
+                ', '.join(engine.table_names())
+            ))
+        elif BIND.startswith('sqlite'):
+            for table in engine.table_names():
+                conn.execute("DELETE FROM %s" % table)
 
         trans.commit()
         conn.close()
